Implemented attribute-map builder
There is now a remote-directory that contains snippits of xml to build
an attribute-map.xml file out of.

The file gets built in a ruby_block instead of using one of the built-in
chef resources because I need to wait until converge-time to get  the
comtents of the snippets.

Todo:
* documentation
* LWRP for attributes?