Made Faye use the correct ssl cerificates.
It'd be nice to be able to use exactly the same cert files for both
Apache and Thin, but they have conflicting opinions about what should be
in each file.

There might be a convergence problem here.  The execute resource doesn't
run if the input files don't exist, but if it doesn't Thin might not
start.  Dunno